Split up Hugging Face Inference API's embeddings to a subclass
